requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
exifread>=3.0.0
folium>=0.15.0
piexif>=1.1.3
pyinstaller>=6.0.0
pyinstaller-hooks-contrib>=2023.8